+++ compositor_pipeline/src/pipeline/webrtc/whip_input.rs
@@ -48,9 +48,7 @@ impl WhipInput {
 
         let bearer_token = options.bearer_token.unwrap_or_else(generate_token);
 
-        let vulkan_available = ctx.graphics_context.has_vulkan_support();
-        let video_preferences =
-            resolve_video_preferences(options.video_preferences, vulkan_available)?;
+        let video_preferences = resolve_video_preferences(options.video_preferences, &ctx)?;
 
         state.inputs.add_input(
             &endpoint_id,
@@ -97,14 +95,15 @@ impl<T> Iterator for AsyncReceiverIter<T> {
 
 fn resolve_video_preferences(
     video_preferences: Vec<WhipVideoDecoderOptions>,
-    vulkan_available: bool,
+    ctx: &Arc<PipelineCtx>,
 ) -> Result<Vec<VideoDecoderOptions>, InputInitError> {
+    let vulkan_supported = ctx.graphics_context.has_vulkan_support();
     let video_preferences: Vec<VideoDecoderOptions> = video_preferences
         .into_iter()
         .flat_map(|preference| match preference {
             WhipVideoDecoderOptions::FfmpegH264 => vec![VideoDecoderOptions::FfmpegH264],
             WhipVideoDecoderOptions::VulkanH264 => {
-                if vulkan_available {
+                if vulkan_supported {
                     vec![VideoDecoderOptions::VulkanH264]
                 } else {
                     warn!("Vulkan is not supported, skipping \"vulkan_h264\" preference");
@@ -117,7 +116,7 @@ fn resolve_video_preferences(
                 vec![
                     VideoDecoderOptions::FfmpegVp9,
                     VideoDecoderOptions::FfmpegVp8,
-                    if vulkan_available {
+                    if vulkan_supported {
                         VideoDecoderOptions::VulkanH264
                     } else {
                         VideoDecoderOptions::FfmpegH264
